requirements.txt
RPi.GPIO==0.6.3
argparse==1.2.1
google-api-python-client==1.6.1
httplib2==0.9.2
oauth2client==4.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
wsgiref==0.1.2
